US138 T142 Reorganize test folder structure
