+++ Assets/Scripts/Managers/PlayerManager.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using UnityEngine;
