+++ Assets/Tests/EditMode/BoardManagerTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f2e513a0bdf3d44ba04368e01a48e78
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
