+++ Assets/Tests/EditMode/EventsTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a98fc998f18ea00439a1315375040189
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
