+++ Assets/Tests/EditMode/GameManagerTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b64b72ec997fef349b957eee8bf729c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
