--- Assets/Tests/EditMode/PlayerManagerTestBase.cs
@@ -1,25 +1,0 @@
-using NUnit.Framework;
-using UnityEngine;
-
-public class PlayerManagerTestBase : ManagerTestBase
-{
-    protected GameObject gameObject;
-    protected PlayerManager playerManager;
-    
-    [SetUp]
-    public virtual void SetUp()
-    {
-        gameObject = new GameObject("PlayerManagerTests");
-        playerManager = gameObject.AddComponent<PlayerManager>();
-
-        playerManager.playerAddedEventChannel = CreateChannel<PlayerEventChannel>();
-        playerManager.playerRemovedEventChannel = CreateChannel<PlayerEventChannel>();
-        playerManager.initializePlayerCountChannel = CreateChannel<IntEventChannel>();
-    }
-
-    [TearDown]
-    public virtual void TearDown()
-    {
-        DestroyTestObjects(gameObject);
-    }
-}
