+++ Assets/Tests/EditMode/PlayerManagerTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bf78b19a5e7352a46a7c8637c08583d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
