+++ Assets/Tests/EditMode/PlayerTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54988a6fa66a43c4fb2e92f58bc97e73
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
