+++ Assets/Tests/EditMode/SpacesTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 647bd0d63c058084792bb89e07b5b9d6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
