+++ Assets/Tests/EditMode/_TestFrameworkVerification.cs
@@ -1,5 +1,5 @@
-using UnityEngine;
 using NUnit.Framework;
+using UnityEngine;
 
 namespace Tests.EditMode
 {
